deleting albums now also deletes the associated images

+++ backend/src/main/java/com/example/backend/service/AlbumService.java
@@ -6,7 +6,11 @@ import com.example.backend.api.model.Album;
 import com.example.backend.api.model.AlbumImages;
 import com.example.backend.api.model.Photo;
 import com.example.backend.exception.AlbumNotFoundException;
+import com.example.backend.exception.DeletePhotoException;
 import com.example.backend.exception.PhotoNotFoundException;
+import jakarta.transaction.Transactional;
+import org.hibernate.sql.Delete;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -16,10 +20,12 @@ public class AlbumService {
 
     private final PhotoRepository photoRepository;
     private final AlbumRepository albumRepository;
+    private final PhotoService photoService;
 
-    public AlbumService(PhotoRepository photoRepository, AlbumRepository albumRepository) {
+    public AlbumService(PhotoRepository photoRepository, AlbumRepository albumRepository, PhotoService photoService) {
         this.photoRepository = photoRepository;
         this.albumRepository = albumRepository;
+        this.photoService = photoService;
     }
 
     /**
@@ -50,17 +56,24 @@ public class AlbumService {
         return album.getId();
     }
 
-    /**
-     * Delete an album by its unique identifier.
-     * @param id the unique identifier of the album to delete
-     * @throws AlbumNotFoundException
-     */
+    @Transactional
     public void deleteAlbum(String id) throws AlbumNotFoundException {
         Album album = albumRepository.findAlbumById(id);
+        System.out.println("Attempting to delete album with id: " + id);
+        List<Integer> photoIds = photoRepository.findPhotoIdsByAlbumId(id);
+        System.out.println("Test");
+        System.out.println("Found " + photoIds.size() + " photos in album with id: " + id);
 
         if (album == null) throw new AlbumNotFoundException("Album with id " + id + " not found");
 
+        if (photoIds.isEmpty()) return;
+        try {
+            photoService.deletePhotos(photoIds);
+        } catch (Exception e) {
+            throw new RuntimeException("Failed to delete photos in album: " + e.getMessage());
+        }
         albumRepository.delete(album);
+
     }
 
     /**
